% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multi_varfun.R
\name{multi_varfun}
\alias{multi_varfun}
\title{Multilevel Variance Function Regression}
\usage{
multi_varfun(outcome, indep = 1, level2, level3 = 1, rslope2 = 1,
  rslope3 = 1, data)
}
\arguments{
\item{outcome}{Outcome variable.}

\item{indep}{List of independent variables (e.g., "age+sex+edu+gdp").}

\item{level2}{ID-Variable for level 2 units (e.g., countryID).}

\item{level3}{ID-Variable for level 3 units (e.g., countryID). Defaults to 1 denoting estimation of 2 level model.}

\item{rslope2}{List of variables for which random slopes are estimated for level 2 (e.g., "edu+age"). Defaults to 1 denoting no random slope estimation.}

\item{rslope3}{List of variables for which random slopes are estimated for level 3 (e.g., "edu+age"). Defaults to 1 denoting no random slope estimation.}

\item{data}{Name of dataset.}
}
\description{
This package run multilevel variance function regression.
}
\examples{
## Ethnic disparities in math competences:
ethn_math <- multi_varfun(outcome="math", indep="migrant+edu_f+age+grade+sex", level2="cid", rslope="migrant", data=PISA)
summary(ethn_math$meanmodel)
summary(ethn_math$varmod)
}
\author{
Christoph Spörlein, \email{christoph.spoerlein@uni-bamberg.de}
}
\references{
Western and Bloome 2009: Variance Function Regression for studying Inequality. Sociological Methodology 39:293-326.
}
\keyword{function}
\keyword{multilevel}
\keyword{regression}
\keyword{variance}

